+++ src/classes/Node.ts
@@ -8,20 +8,20 @@ import {
   PeerDetails,
   PublicKey,
   Txid,
-} from './Bindings';
+} from "./Bindings";
 import {
   addressToString,
   createChannelDetailsObject,
   createPaymentDetails,
   createPeerDetailsObject,
   stringToAddress,
-} from '../utils';
+} from "../utils";
 
-import { NativeLoader } from './NativeLoader';
-import { ChannelConfig } from './ChannelConfig';
+import { NativeLoader } from "./NativeLoader";
+import { ChannelConfig } from "./ChannelConfig";
 
 export class Node extends NativeLoader {
-  id: string = '';
+  id: string = "";
 
   /**
    * Create node id
@@ -75,7 +75,9 @@ export class Node extends NativeLoader {
    */
   async listeningAddresses(): Promise<Array<NetAddress> | null> {
     let addresses = await this._ldk.listeningAddresses(this.id);
-    return addresses == undefined ? null : addresses.map((i) => stringToAddress(i));
+    return addresses == undefined
+      ? null
+      : addresses.map((i) => stringToAddress(i));
   }
 
   /**
@@ -93,8 +95,15 @@ export class Node extends NativeLoader {
    * @requires [amountMsat] amount in milli sats
    * @returns {Promise<TxId>}
    */
-  async sendToOnchainAddress(address: Address, amountMsat: number): Promise<Txid> {
-    let txid = await this._ldk.sendToOnchainAddress(this.id, address.addressHex, amountMsat);
+  async sendToOnchainAddress(
+    address: Address,
+    amountMsat: number
+  ): Promise<Txid> {
+    let txid = await this._ldk.sendToOnchainAddress(
+      this.id,
+      address.addressHex,
+      amountMsat
+    );
     return new Txid(txid);
   }
 
@@ -104,7 +113,10 @@ export class Node extends NativeLoader {
    * @returns {Promise<TxId>}
    */
   async sendAllToOnchainAddress(address: Address): Promise<Txid> {
-    let txid = await this._ldk.sendAllToOnchainAddress(this.id, address.addressHex);
+    let txid = await this._ldk.sendAllToOnchainAddress(
+      this.id,
+      address.addressHex
+    );
     return new Txid(txid);
   }
 
@@ -134,8 +146,17 @@ export class Node extends NativeLoader {
    * @requires [persist] open node permanently or not
    * @returns {Promise<boolean>}
    */
-  async connect(nodeId: string, address: NetAddress, persist: boolean): Promise<boolean> {
-    return await this._ldk.connect(this.id, nodeId, addressToString(address), persist);
+  async connect(
+    nodeId: string,
+    address: NetAddress,
+    persist: boolean
+  ): Promise<boolean> {
+    return await this._ldk.connect(
+      this.id,
+      nodeId,
+      addressToString(address),
+      persist
+    );
   }
 
   /**
@@ -184,8 +205,15 @@ export class Node extends NativeLoader {
    * @requires [counterpartyNodeId] publicKey of counterparty Node
    * @returns {Promise<boolean>}
    */
-  async closeChannel(channelId: ChannelId, counterpartyNodeId: PublicKey): Promise<boolean> {
-    return await this._ldk.closeChannel(this.id, channelId.channelIdHex, counterpartyNodeId.keyHex);
+  async closeChannel(
+    channelId: ChannelId,
+    counterpartyNodeId: PublicKey
+  ): Promise<boolean> {
+    return await this._ldk.closeChannel(
+      this.id,
+      channelId.channelIdHex,
+      counterpartyNodeId.keyHex
+    );
   }
 
   /**
@@ -209,8 +237,15 @@ export class Node extends NativeLoader {
    * @requires [amountMsat]
    * @returns {Promise<PaymentHash>}
    */
-  async sendPaymentUsingAmount(invoice: string, amountMsat: number): Promise<PaymentHash> {
-    let hash = await this._ldk.sendPaymentUsingAmount(this.id, invoice, amountMsat);
+  async sendPaymentUsingAmount(
+    invoice: string,
+    amountMsat: number
+  ): Promise<PaymentHash> {
+    let hash = await this._ldk.sendPaymentUsingAmount(
+      this.id,
+      invoice,
+      amountMsat
+    );
     return new PaymentHash(hash);
   }
 
@@ -220,8 +255,15 @@ export class Node extends NativeLoader {
    * @requires [amountMsat]
    * @returns {Promise<PaymentHash>}
    */
-  async sendSpontaneousPayment(amountMsat: number, nodeId: PublicKey): Promise<PaymentHash> {
-    let hash = await this._ldk.sendSpontaneousPayment(this.id, amountMsat, nodeId.keyHex);
+  async sendSpontaneousPayment(
+    amountMsat: number,
+    nodeId: PublicKey
+  ): Promise<PaymentHash> {
+    let hash = await this._ldk.sendSpontaneousPayment(
+      this.id,
+      amountMsat,
+      nodeId.keyHex
+    );
     return new PaymentHash(hash);
   }
 
@@ -232,8 +274,17 @@ export class Node extends NativeLoader {
    * @requires [expirySecs] number
    * @returns {Promise<boolean>}
    */
-  async receivePayment(amountMsat: number, description: string, expirySecs: number): Promise<string> {
-    return await this._ldk.receivePayment(this.id, amountMsat, description, expirySecs);
+  async receivePayment(
+    amountMsat: number,
+    description: string,
+    expirySecs: number
+  ): Promise<string> {
+    return await this._ldk.receivePayment(
+      this.id,
+      amountMsat,
+      description,
+      expirySecs
+    );
   }
 
   /**
@@ -242,8 +293,15 @@ export class Node extends NativeLoader {
    * @requires [expirySecs] number
    * @returns {Promise<boolean>}
    */
-  async receiveVariableAmountPayment(description: string, expirySecs: number): Promise<string> {
-    return await this._ldk.receiveVariableAmountPayment(this.id, description, expirySecs);
+  async receiveVariableAmountPayment(
+    description: string,
+    expirySecs: number
+  ): Promise<string> {
+    return await this._ldk.receiveVariableAmountPayment(
+      this.id,
+      description,
+      expirySecs
+    );
   }
 
   /**
@@ -281,7 +339,9 @@ export class Node extends NativeLoader {
    * @returns {Promise<PaymentDetails>}
    */
   async payment(paymetHash: PaymentHash): Promise<PaymentDetails> {
-    return createPaymentDetails(await this._ldk.payment(this.id, paymetHash.field0));
+    return createPaymentDetails(
+      await this._ldk.payment(this.id, paymetHash.field0)
+    );
   }
 
   /**
@@ -316,7 +376,11 @@ export class Node extends NativeLoader {
    * @requires [pkey] public key of node
    * @returns {Promise<string>}
    */
-  async verifySignature(msg: Array<number>, sig: string, pkey: PublicKey): Promise<boolean> {
+  async verifySignature(
+    msg: Array<number>,
+    sig: string,
+    pkey: PublicKey
+  ): Promise<boolean> {
     return await this._ldk.verifySignature(this.id, msg, sig, pkey.keyHex);
   }
 
@@ -338,6 +402,78 @@ export class Node extends NativeLoader {
       channelId.channelIdHex,
       counterpartyNodeId.keyHex,
       channelConfig.id
+    );
+  }
+
+  /**
+   * Returns whether the [`Node`] is running.
+   *
+   * @returns {Promise<boolean>}
+   */
+  async isRunning(): Promise<boolean> {
+    return await this._ldk.isRunning(this.id);
+  }
+
+  /**
+   * Sends payment probes over all paths of a route that would be used to pay the given invoice.
+   *
+   * This may be used to send "pre-flight" probes, i.e., to train our scorer before conducting
+   * the actual payment. Note this is only useful if there likely is sufficient time for the
+   * probe to settle before sending out the actual payment, e.g., when waiting for user
+   * confirmation in a wallet UI.
+   *
+   *
+   * Otherwise, there is a chance the probe could take up some liquidity needed to complete the
+   * actual payment. Users should therefore be cautious and might avoid sending probes if
+   * liquidity is scarce and/or they don't expect the probe to return before they send the
+   * payment. To mitigate this issue, channels with available liquidity less than the required
+   * amount times [`Config::probing_liquidity_limit_multiplier`] won't be used to send
+   * pre-flight probes.
+   *
+   * @requires [invoice]
+   * @returns {Promise<PaymentHash>}
+   */
+  async sendPaymentProbes(invoice: string): Promise<PaymentHash> {
+    let hash = await this._ldk.sendPaymentProbes(this.id, invoice);
+    return new PaymentHash(hash);
+  }
+
+  /**
+   * Sends payment probes over all paths of a route that would be used to pay the given
+   * zero-value invoice using the given amount.
+   *
+   * This can be used to send pre-flight probes for a so-called "zero-amount" invoice, i.e., an
+   * invoice that leaves the amount paid to be determined by the user.
+   *
+   * @requires [invoice]
+   * @requires [amountMsat]
+   * @returns {Promise<boolean>}
+   */
+  async sendPaymentProbesUsingAmount(
+    invoice: string,
+    amountMsat: number
+  ): Promise<boolean> {
+    return await this._ldk.sendPaymentProbesUsingAmount(
+      this.id,
+      invoice,
+      amountMsat
+    );
+  }
+
+  /**
+   * Sends payment probes over all paths of a route that would be used to pay the given amount to the given `node_id`.
+   * @requires [invoice]
+   * @requires [amountMsat]
+   * @returns {Promise<boolean>}
+   */
+  async sendSpontaneousPaymentProbes(
+    amountMsat: number,
+    nodeId: PublicKey
+  ): Promise<boolean> {
+    return await this._ldk.sendSpontaneousPaymentProbes(
+      this.id,
+      amountMsat,
+      nodeId.keyHex
     );
   }
 }